# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SocialApp — unified social network + messaging platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /follows/{userId}:
    post:
      tags: [Follows]
      operationId: followUser
      summary: Follow a user
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Followed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowResult'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Follows]
      operationId: unfollowUser
      summary: Unfollow a user
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Unfollowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowResult'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  # ── Posts ─────────────────────────────────────────────────────────────────
...
  /posts/{postId}/comments:
    get:
      tags: [Comments]
      operationId: getComments
      summary: Get comments on a post
      parameters:
        - name: postId
          in: path
          required: true
          schema: { type: string }
        - name: cursor
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Comments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentListPage'
    post:
      tags: [Comments]
      operationId: createComment
      summary: Comment on a post
      parameters:
        - name: postId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentInput'
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
  /comments/{commentId}:
    patch:
      tags: [Comments]
      operationId: updateComment
      summary: Edit a comment
      parameters:
        - name: commentId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentUpdate'
      responses:
        '200':
          description: Updated comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
    delete:
      tags: [Comments]
      operationId: deleteComment
      summary: Delete a comment
      parameters:
        - name: commentId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Deleted
  /comments/{commentId}/replies:
    get:
      tags: [Comments]
      operationId: getReplies
      summary: Get replies to a comment
      parameters:
        - name: commentId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Replies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentListPage'
    post:
      tags: [Comments]
      operationId: createReply
      summary: Reply to a comment
      parameters:
        - name: commentId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentInput'
      responses:
        '201':
          description: Reply created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'

  # ── Stories ───────────────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    UserSummary:
      type: object
      required: [id, username, displayName]
      properties:
        id: { type: string }
        username: { type: string }
        displayName: { type: string }
        avatarUrl: { type: ["string", "null"] }
        isFollowing: { type: boolean }
...
    FollowResult:
      type: object
      required: [following, followersCount]
      properties:
        following: { type: boolean }
        followersCount: { type: integer }

    # ── Posts ────────────────────────────────────────────────────────────────
...
    Comment:
      type: object
      required: [id, author, content, repliesCount, createdAt]
      properties:
        id: { type: string }
        author:
          $ref: '#/components/schemas/UserSummary'
        content: { type: string }
        postId: { type: ["string", "null"] }
        parentId: { type: ["string", "null"] }
        repliesCount: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CommentInput:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1 }
    CommentUpdate:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1 }
    CommentListPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        nextCursor: { type: ["string", "null"] }

    # ── Stories ──────────────────────────────────────────────────────────────
...